History Bill Done - print bill done - save login done

+++ lib/FireBase/FireStore.dart
@@ -8,42 +8,46 @@ class FireStoreUser{
       'price':price,
       'qrCode':qrCode
     };
-    Firestore.instance.collection('Stores').document(userId).collection('product').where("name",isEqualTo: name).getDocuments()
-    .then((QuerySnapshot docs){
-      print("ten giong la : ${docs.documents.length}");
-      if(docs.documents.length == 0){
-        if(qrCode != ""){
-          Firestore.instance.collection('Stores').document(userId).collection('product').where("qrCode",isEqualTo: qrCode).getDocuments()
-              .then((QuerySnapshot docs)  {
-            print(docs.documents.length);
-            if(docs.documents.length == 0){
-              // them mat hang len db
-              _db.collection("Stores").document(userId).collection("product").document().setData(product).then((val){
-                print("add product ok");
-              }).catchError((err){
-                print("co loi nang : ${err}");
-              });
-              onSuccess("Tạo mặt hàng trên hệ thống thành công");
-            }else{
-              onErr("Đã có mặt hàng trùng mã vạch với sản phẩm này trên hệ thống");
-              print("da co tren db");
+    if(name.length > 20){
+      onErr("Tên sản phẩm phải dưới 20 kí tự");
+    }else{
+      Firestore.instance.collection('Stores').document(userId).collection('product').where("name",isEqualTo: name).getDocuments()
+          .then((QuerySnapshot docs){
+        print("ten giong la : ${docs.documents.length}");
+        if(docs.documents.length == 0){
+          if(qrCode != ""){
+            Firestore.instance.collection('Stores').document(userId).collection('product').where("qrCode",isEqualTo: qrCode).getDocuments()
+                .then((QuerySnapshot docs)  {
+              print(docs.documents.length);
+              if(docs.documents.length == 0){
+                // them mat hang len db
+                _db.collection("Stores").document(userId).collection("product").document().setData(product).then((val){
+                  print("add product ok");
+                }).catchError((err){
+                  print("co loi nang : ${err}");
+                });
+                onSuccess("Tạo mặt hàng trên hệ thống thành công");
+              }else{
+                onErr("Đã có mặt hàng trùng mã vạch với sản phẩm này trên hệ thống");
+                print("da co tren db");
 
-            }
-          });
+              }
+            });
+          }else{
+            _db.collection("Stores").document(userId).collection("product").document().setData(product).then((val){
+              print("add product ok");
+            }).catchError((err){
+              print("co loi nang : ${err}");
+            });
+            onSuccess("Tạo mặt hàng trên hệ thống thành công");
+          }
         }else{
-          _db.collection("Stores").document(userId).collection("product").document().setData(product).then((val){
-            print("add product ok");
-          }).catchError((err){
-            print("co loi nang : ${err}");
-          });
-          onSuccess("Tạo mặt hàng trên hệ thống thành công");
-        }
-      }else{
-        print("da co tren db");
-        onErr("Đã có mặt hàng trùng tên này trong cửa hàng của bạn");
+          print("da co tren db");
+          onErr("Đã có mặt hàng trùng tên này trong cửa hàng của bạn");
 
-      }
-    });
+        }
+      });
+    }
   }
   void deleteProduct({String userId,String documentID,Function(String) onSuccess,Function(String) onErr}){
     _db.collection('Stores').document(userId).collection('product').document(documentID).delete()
@@ -65,8 +69,6 @@ class FireStoreUser{
     _db.collection('Stores').document(userId).collection('product').document(documentID).updateData(product)
     .then((val){
       onSuccess("sửa dữ liệu thành công!");
-    }).catchError((err){
-      onErr("lỗi ${err}");
     });
   }
   void checkStore({String userId}){
@@ -118,12 +120,21 @@ class FireStoreUser{
   void ChangeQuantityProductBill({String userId,String docId,int quantityUpdate}){
     Firestore.instance.collection('Stores').document(userId).collection('currentBill').document(docId).updateData({ 'quantity': quantityUpdate });
   }
+  void UpdateQuantityProductStore({String userId,String docId,int quantityUpdate}){
+    Firestore.instance.collection('Stores').document(userId).collection('product').document(docId).updateData({ 'quantity': quantityUpdate });
+
+  }
   void CancelBill({String userId,Function(String) onSuccess,Function(String) onErr}){
     _db.collection('Stores').document(userId).collection('currentBill').getDocuments().then((snapshot){
-      for (DocumentSnapshot ds in snapshot.documents){
-        ds.reference.delete();
+      if(snapshot.documents.length>0){
+        for (DocumentSnapshot ds in snapshot.documents){
+          ds.reference.delete();
+        }
+        onSuccess("Xóa hóa đơn thành công!");
+      }else{
+        onErr("Hóa đơn không có mặt hàng nào để xóa");
       }
-      onSuccess("Xóa hóa đơn thành công!");
+
 
     });
   }
@@ -134,8 +145,17 @@ class FireStoreUser{
       var product=[];
 
       if(snapshot.documents.length>0){
-        print(snapshot.documents.length);
-        snapshot.documents.map((DocumentSnapshot document){
+        print("so l sp hien tai ${snapshot.documents.length}");
+        snapshot.documents.map((DocumentSnapshot document)  {
+          checkProductInStore(userId: userId,qrCode: document['product']['qrCode'],name: document['product']['name']
+              ,onErr: (txt){}
+              ,onS: (data){
+                int quatityUpdate = data['quantity'] - document['quantity'];
+                UpdateQuantityProductStore(docId: data.documentID,userId: userId,quantityUpdate: quatityUpdate);
+
+              }
+          );
+
           product.add(document.data);
         }).toList();
         print(product);
@@ -144,10 +164,40 @@ class FireStoreUser{
           'bill':product,
           'total':total
         };
+
         _db.collection('Stores').document(userId).collection('BillsHistory').document().setData(bill);
         onSuccess("Tính tiền thành công!");
 
+
+
+
+      }else{
+        onErr("Chưa có sản phẩm nào để tính tiền");
       }
     });
   }
+  void checkProductInStore({String qrCode,String userId,String name,Function(String) onErr(String),Function(DocumentSnapshot)onS}){
+    _db.collection('Stores').document(userId).collection('product').where("qrCode",isEqualTo: qrCode).getDocuments()
+        .then((QuerySnapshot docs){
+      if(docs.documents.length != 0){
+         docs.documents.map((DocumentSnapshot document){
+          print(document.data);
+          onS(document);
+        }).toList();
+      }else{
+        _db.collection('Stores').document(userId).collection('product').where("name",isEqualTo: name).getDocuments()
+            .then((QuerySnapshot docs)  {
+          if(docs.documents.length != 0){
+              docs.documents.map((DocumentSnapshot document){
+                print(document.data);
+                onS(document);
+            }).toList();
+          }else{
+            onErr("Lỗi!");
+          }
+        });
+      }
+    });
+  }
+
 }
